Reviewed and corrected synchronization of access to DB configuration data. JIRA: CXX-6137
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@68168 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/app/netcache/nc_storage.hpp
@@ -105,7 +105,7 @@ public:
     static Uint4 GetNewBlobId(void);
 
     /// Get number of files in the database
-    static int GetNDBFiles(void);
+    static size_t GetNDBFiles(void);
     /// Get total size of database for the storage
     static Int8 GetDBSize(void);
     static Int8 GetDiskFree(void);
